Hack to make claims ordering explicit
It orders by date and id.  The factory creates all the records
simultaneously, so they wind up ordered by id.  Not sure this is desired
behaviour, but I'm putting this in now to make the spec pass.  As
ordering is important for the vairous tables, I will have a discussion
with Stuart et al. and then put some specific specs in place regarding
the desired ordering.