Detailed financial summary of advocate claims
A financial summary of authorised and outstanding claims with a link through to a detailed list
showing what claims make up that summary.

the summary and list is constructed depending on the context it's viewed
in. I.e either an as an advocate or as an admin of a chamber.

- Adds controllers and views for detailed financial summary
- renames Claims::Summary to Claims::FinancialSummary and the variable
  parent to context within that class and the claims controller
- fixes and documents BigDecimal bug when summing the financial summary
- improves financial summary cukes to test for value
- fixes advocate dashboard failure when advocate admin does not have a chamber