basic api for draft claim submission
This is the initial work on production of an API
for the submission of draft claims. The intent is to provide
a means by which Case Management Systems used by many big
advocate chambers can submit claims to ADP programmatically,
thereby removing the need for rekeying of data.

This branch includes:

  * use of Grape gem for quick API
  * use of Swagger gem for automated documentation
  * use of Swagger UI for a testing interface for the API
  * a claim creation API endpoint
  * a claim validation API endpoint
  * error handling for the above endpoints
  * rspecs for the above endpoints