fix pagination bug
There was a bug whereby if on page 2 of case worker
archived claims when you hit Your claims it
tries to go to page 2 of Your claims (which may not exist).

This fix exludes the page param from the nav bar links.

also, modified search summary to take account of pagination
information.