fix broken archive search on gamma
ON gamma most archived claims are in the "archived pending delete” state - this plus the fact that search does not specify a list of states to filter by means the default dashboard_displayable_states is used to filter claims. This default does NOT include archived_pending_delete, hence search returns 0 found for most searches (you can search and find claims in any other state on the dashboard).