Enable STI on claims
This PR turns adds Single Table Inheritance to the Claim model, renaming it Claim::BaseClaim.  This is
in preparation for creating a Claim::AdvocateClaim and Claim::LitigatorClaim model.

In addition to the renaming and migration of the claims table, changes were made to the Assessment model to ensure that
only one can ever exist for a claim - see note at top of Assessment model for details on how to create and update assessments.