Fix Migration Bug
Migrations need to work both when run as part of db:reload and running all migrations on an empty database and when migrating an existing database on live.

This was failing because it relied on the fee_categories to populate the type on the FeeType record, which didn't exist when migrating an empty database.

this fixed that issue