Write Management Information Reports to database
The original method of writing management information reports as a local CSV file on the web server does not work
in a clustered server environment because requests can't be guaranteed to go to one particular server.

This pull requests changes the Management Information report generation and retrieval to use the database instead.