Setting allocation type set on submit
The allocation type is set whenever a TransferClaim transitions to submitted state.
This is acheived by a after_transmission hook which calls set_allocation_type on the claim.
This method is implemented as an empty method on BaseClaim, but on TransferClaim it
calls TransferBrain to work out the allocation type according tothe fields on the TransferDetail and
save it to the claim record.