Rake task to audit attached documents
There have been several reports of "corrupted" documents.  It turns out that the
files don't actually exist in the S3 container.  All the ones investigated so far
were in a particular directory.

This rake task will examine all documents on the system, and report whether or not they are
present, and when they were created, to see if we can see some sort of correlation of
missing documents.