Verify Saving Documents
There have been instances of attached documents not being in the S3 storage later.  We
don't know whether they were successfully uploaded and then subsequently dissappeared (unlikely)
or whether they simply were not successfully uploaded correctly in the first place.

This PR adds a verification stage to the document upload, which ensures that we can read back the
document once it's been saved, and records this fact on the document model.

It also adds additional logging around document saving.