WIP statistics collector and store
This PR imnplements the first stage of producing a dashboard to track the daily number of claims that are being
submitted.  It models a statistics table to store the results, and the first collector, to store the number of submitted claims.

It does not yet implement the cron job to record this information daily, nor the gecko API.