Fix cause of flickering cukes
Cukes were occasionally flickering because the offence class "E: Burglary" couldn't
be found.  This was caused the cukes which were marked as no-seeding running before
the ones marked as requiring seeding, and creating a random offence class.  When the
cukes that required seeding ran later, they don't overwrite existing offence classes
in the database.  The cuke flickered when the randomly created offence class was E,
with a different name.

The solution was to require seeding for all cukes.  This does not add to the running
time of the cucumber suite, because it is only ever carried out once.