Use working days for time to completion stat report
Previously, the time at which the claim was submitted was subtracted from the time
it was decided giving a number of seconds which were totalled for all claims
decided on the day in question, and then converted to an average number of days.

Now, the time in whole working days is calculated between submission and deision,
and an average worked out from that.