Disallow lowercase letters in supplier numbers and better exception handling.
Make sure all current supplier numbers in database are uppercase and do not
allow future ones to contain lower case letters.

Also, a not handled exception could be raised when allocating claims with
validations not met (like in this case supplier numbers). Now we handle
the exception and gives the user feedback (no more 500 error page).