New ActionMailer delivery method for GOV.UK Notify service.
This PR prepare the work to be able to send emails via an external
service (GOV.UK Notify) maintaining the same interface as any other
standard ActionMailer.
Also it will use ActiveJob, and the standard delivery configuration
for enqueued deliveries.

The way it is done will simplify extracting to a gem as there is no
coupling with our own app.

An example 'NotifyMailer' is included with a test email method.