Remove the option 'Travel costs' from the disbursements options.
As there are claims using this disbursement and we can't reassign those
to another disbursement, a temporal workaround was discussed to 'hide'
the option from the dropdowns and API but to continue in the database so
old claims will continue working and validating, but new ones will not
be able to select 'Travel costs'.

Sometime from now it will be safe to remove the option from the database.

PT#126855639