Open feedback and bug report links in separate tabs.
Also if we detect they came from a claims page we add a note to indicate
that the claim is still available in another tab.

This is however not 100% reliable due to how referrer works, but should
catch most of the important cases (giving feedback while creating a claim).

PT#127165105