Do not allow deleting case workers with sent messages.
This is a safety mechanism to ensure no case workers can be deleted
if they have created any message in any claim, so no claims can be
rendered orphaned, and thus not possible to process.

A soft-delete will be a better way of handle this situations, and will
be implemented in the coming weeks.