Enabled test suite to be run in any time zone
Rails stores all dates on the database in UTC and converts them to the Time
Zone specified in the configuration, in our case London.

If tests are written comparing various DB dates to Time.now, or Time.now - 2.days
these will fail if the machine running the tests is in any timezone other than
London.  These fix that by making the tests use Time.zone.now, which will
use the rails config time.