Offence seletion for Litigiators to mimic that for advocates (#1705)
* Offence seletion for Litigiators to mimic that for advocates

Previsouly, litigators just selected the offence class, and an offence category of miscellaneous/other
relating to that class was automatically selected.

Now, the process has been changed to be exactly the same as for advocates: i.e. the category is selected
and then the class, (or auto selected if there is only one class for the selected category).  The valiation
for misc/other category for LGFS claims has been removed.

* Update API documentation and release notes