Add deleted_at to User and CaseWorker
This PR sets up the framework for soft delete of caseworkers and users.

It adds a 'deleted_at' field to CaseWorker and User models, and sets the
default scope to exclude any records where deleted at is not null.  Also
adds an including_deleted scope which returns all records.

This has no effect on the current deletion process as that hasn't yet been
changed to use the deleted_at fields.