Backend ability to clone details from one claim to another.
Only defendants with representation orders, and court_id are cloned at the moment
due to any other details being very dependant on the schema and case type, not being
possible at this point to make assumptions.

Once the UI is implemented and we have a clear understanding of how this 'cloning' is
going to be performed, some other details might be transfered.