Soft deletable concern for CaseWorker and User
Caseworkers are no longer deleted from the database because of claims and messages that
may reference the deleted caseworker. Now a new 'deleted_at' field on CaseWorker and User
is given a timestamp, and any record where that field is not null is considered to be deleted.

A new scope active on CaseWorker and User will only return records where deleted_at is nil,
and Caseworker queries have been updated to use that scope where approperiate.

This is implemented as a Concern as it may well be used on Claims as well.