Claim scheme filters in the main claim list for external users.
The filters will only appear when the user has the ability to create AGFS
and LGFS claims, otherwise it doesn't make sense to show the filters.

When LGFS filter is selected, the JSON importer is hidden, as we don't
support yet LGFS claim imports and it creates a confusing experience.