Handle multiple error messages on gov_uk_date_felds
There is an issue when an invalid date is input through the form to a gov_uk_date_field
whereby the gov_uk_date_field gem adds an 'Invalid date' message to the errors hash, and
the model validates the field (which is nil) and adds a "Can't be blank" error message
resulting in two error messages.

This PR uses a version of gov_uk_date_field which allows us to control the behaviour
when there are duplicate error messages.

In addition, we also check in the app whether or not there are already messages on a
gov_uk_date_field before validating presence - this is because we have no way of knowing which
validation will trigger first - the model's validations or the gem's validations.

xx