Swagger update.
A bug in swagger gem makes the listing of namespaced endpoints wrong, so to
minimize the impact now all endpoints are arranged inside a generic 'api' header
that can be expanded to show all endpoints.

The bug also forbids us from renaming this 'api' header to something more
meaningful, unfortunately.

https://github.com/ruby-grape/grape-swagger/issues/414