Fix written reasons behaviour.
Previously, after a case worker gave written reasons, the claim would go
back to under certain circunstances to allocated or unallocated state and
also it might remains in awaiting written reasons state.

Now it is consisten and once written reasons are provided, the claim will go
back to the previous state it was before asking for written reasons.

A bit of refactor done extracting complex conditions to helper methods.