Display correct assessment date
An blank assessment is created at the same time a claim is created, and
therefore has a created_at date the same as the claim's created at date.

Proviously, when an assessment was updated with the assessed amounts, the
created at date was also altered (this is a bad hack).

Since the implementation of the CaseWorkerClaimUpdater service, this has
not been happening, so the assessed date is being shown as the claim creation date.

This PR fixes that by working out what the real asssessed date is: the updated_at on an
assessment or the created_at date on the last of any redeterminations.