Store the author user and subject user on transition changes.
Whenever possible, store the ID of the user triggering a state change (author_id),
and the ID of the user receiving the action (subject_id).

For example for allocations, it will store the id of the user triggering the allocation action,
and the id of the user who was allocated said claim.

For transitions not involving another user, only the author_id will be filled (i.e. rejections)