Spike to implement a backend-agnostic caching wrapper.
This will allow us to cache API requests, using as unique keys the endpoints (urls)
including parameters. Future improvement involved filtering out some parameters to
made the hit/miss ratio better (i.e. api_key should be filtered out).

As an example, there is a Redis caching and a Memory caching. Swapping between them
is done via an initialization file.

A caching policy can be provided to enable the caching to stale after N mins, for example.

To test it, run the API smoke tests. The dropdowns will be cached. Run it several times
with Redis server started to see the caching in action.
With Memory caching, each run will wipe the cache so the hit/miss is worse.