Add Unique code to Fee Types table
The Fee types table has a code, but it is not unique, being merely
the code copies from the original forms.

This PR adds a new field - unique_code which can be used to positively
identify a fee type.

Once the migration is done, run rake data:migrate:fee_type_unique_code to
populate the data.

Next steps not included in this PR is:
 - a migration to force not-null integrity on the new column
 - expose unique code in the API
 - allow the API to use unique code as an alternative to database id