UI: Mobile claims list missing sort links
Issue
The responsive tables where set to start as a 1024px break point. This seems to high as there are desktop users that fall into this breakpoint

***What does PR do?***
Adds a responsive row into claims lists that contain the sort links on mobile views

How has this been tested?
Manual browser testing