CCCD to CCR Expense type adapter
This PR provides a new class which will return the LAA bill type and sub type for
a given fee.  It is not yet wired in to provide the LAA bill type and subtype when outputting a claim.

It also changes the fee_type_code on the case type to use the unique code.  This is used to determine the
bill type and sub type for basic fees.