Combine rejected refused
Fix test to reflect combination of rejected/refused

The Geckoboard widget to display the percentage of calims authorised,
part_authorised, refused and rejected was displaying the wrong values -
taking the numbers authorised etc. this month and displaying them as a
percentage of the total number of claims ever in the database.

This PR fixes that bug.