Remove n+1 query from claims
There was an after_initialize callback on Claim::BaseClaim that instantiated an
assessment if not already there.  This meant that when finding a collection of
claims, an query was made to get the assessment for each claim in the collection.

This PR replace that with an after_save hook to create a blank assessment when
the claim is created, and rewrite the assessment factory to cater for this.