Add VAT totals for fees, disbursements and expenses on claim records
Prior to this PR, the total VAT for a claim was calculated by adding
up the VAT amounts on the individual expenses and disbursements, and
calculating the VAT from the fees records.

In order to be able to display individual VAT totals on the claim
summary page, we now calculate the VAT totals individually and store on
the claim record.