Removal of VCR for specs
This PR does two things:

- removes the VCR cassettes which were used to test the allocation_controller.
This controller is now tested to ensure that the appropriate services are called.
And those services will be tested to ensure that the appropriate remote API calls
are made.  The CaseWorkersController (the recipient of the API calls) will be tested
to ensure that the correct responses are made to the requests (this will be done in a
 later PR)

- removes the ability to view the case worker pages locally.  There was a configuration
setting to control whether the case worker allocation pages accessesd the database
directly, or used the API to get the list of claims.  This has now been removed; remote access will always be used.

- Update the README to indicate that a second server running on port 3001 needs to be
started in development mode.