Force email confirmation field to lower case
Devise will automatically downcase the email field.  Unless we do the
same for the email  confirmation field, then the the user creation will
fail because the email does not equal the email confirmation.

The devise configuration is set to not check case sensistivity on email,
so the user will be able to log in with any combination of upper or lower case
in theemail.