add rb-readline to gemfile
byebug has a OS library dependency that is, nonentheless, not a
gem dependency. This results in rails server or console start raising
an error

```
Sorry, you can't use byebug without Readline. To solve this, you need to
rebuild Ruby with Readline support. If using Ubuntu, try `sudo apt-get
install libreadline-dev` and then reinstall your Ruby.
```

Easiest fix to prevent the error is to add rb-readline as dev/test
dependency in the Gemfile.