Add CCR claim endpoint entities and tests (#2028)
* Add CCR claim endpoint entities and tests

This is to form the basis of iteration on
the existing CCR Claim endpoint to the point
where it can become a full claim end point.

* Add CCR claim json schema

To be used to represent current state of JSON returned
by CCR claim endpoint.

* add json schema validation test

To validate ccr claim api endpoint responses
a iterable JSON schema has been applied. This
schema should be modified to represent the
expected response structure.

* Remove internal full claim schema

This schema is not used and just confuses
the approach taken, setting targets that may
not turn out to be appropriate. Iterating on ccr_claim_shema
instead.