Add handler for vat_amount being nil
The front end allows caseworkers to submit a nil value for VAT amount
this causes the redetermination presenter to blow up because it
can't add nil to the total.
This prevented the display of the claim entirely.

This ensures the vat amount will add zero instead of nil and allow the
display of the claim