Use explicit app root path in dockerfile
An unknown change has cause the $APPHOME variable
to not be available in the docker file
resulting in the WORKDIR $APPHOME docker
command raise a `cannot normalise nothing`
error.

This fix simply hard codes the app root
path to be used as the current working directory.
But need to talk to ops about underlying cause
and to check for a better fix or check whether this
is appropriate.