Define the submission stages for advocate claims
As a first attempt to move away from numbered stages that have an
implicit meaning, and instead move towards the use of named stages.

NOTE: Kept the management of the steps as is, so it doesn't impact any
other claim submission type, although the idea is to move all of the
submission to a well defined workflow for each (possibly even storing
the stages in the database so a user can resume their submission from
where they left off.