Add ability to run cucumber features through chrome
Mostly for development/test purposes to be able to see the feature being
run in the actual browser.

NOTE: Due to the way some scenarios are defined, some steps fail while
being run in Chrome as some of the steps definition are not supported by
the chromedriver. The most common example is steps using '.trigger
"click"' instead of using '.click'