Replace separate amount and vat with "gross" total
CCLF requires a vat included flag. For expenses
this can be simplified by adding the VAT claimed
to the "net" amount regardless and specifying
that VAT is always included. Thereby if a claimant
puts 0 VAT he is explicitly choosing to not have VAT
included - so CCLF should NOT modify the amount, which is
achieved by supplying vat_included flag as true??