Fix failing spec for case worker claim updater
For some reason, the previous commit made this spec to fail.

Looking closer at the spec as clear that the spec was sending empty
strings for the redetermination fees and that was being used to
initialize the Redetermination model (instead of defaulting to 0.0)

Theses changes ensure any nil or blank values are ignored for the
purpose of saving the redeterminations or the assessments.

Also:

This service seems to mixing up responsabilites having quite a lot of
validation code. That should probably be in a validator class that is
used by the service. The service should mainly be responsabile for
updating the claim.