Change available claim types to include advocate interim claims
Note to the fact, that the advocate interim claim is feature flagged so
it will only be available in environments where the AGFS fee reform
feature is active. This is to allow us to keep on delivering the
features without creating a massive feature branch that ends up in a big
bang deploy.