Increase allocation logging
We are seeing a small number of cases that, despite the error
handling in allocate!, end up in a state where the claim is
allocated, but no link is made to a case worker.

This results in the claim disappearing from the re/allocation
tables and , co-incidentally, MI breaking.
Despite our best efforts, the error handling is not trapping
this, so this increase in logging will hopefully allow us to
track down the circumstances where this occurs.