Add back button to claim forms
By default, if there's no step to go back to, the user will see a 'Home'
button and will be redirected to the list of claims.

Otherwise the user will see a 'Back' button and will be redirected to
the previous step of submission.

Also:

- For ALL claim forms the former navigation menu was removed (to be
aligned with the new designs).