Remove unnecessary display navigation checks
Previous PR introduced a temporary method to check that the primary
navigation should be displayed or not for a given controller action.

With the introduction of the nested layouts, this is no longer necessary
as each controller action can have their own specific layouts (all the
claim form ones now have one).