Preserve current behaviour for LGFS interim checks
Until now, there was only one type of claim that was 'interim' (LGFS
interim claim), the checks for it were all being done using only
'interim?' not taking into account the type of claim (AGFS or LGFS).

With the introduction of the AGFS interim, until we're clear of what
behaviour should apply to this new type of claim, any existent check on
'interim?' only was changed to be 'lgfs? && interim?' to preserve the
existent behaviour.