Set AGFS release date to be at least 3 months old
In order for us to be able to test the whole AGFS interim claim flow as
one of the requirements is for the warrant fee issued date to be at
least 3 months ago and after the earliest representation order date.

NOTE: This value will need to be changed to the correct date once the
changes go live!