Refactor the behaviour to determine eligible basic fee types
Basic fee types are now seeded into the database with additional roles
that are associated with a given scheme (this can potentially be
replaced all together with the introduction of the fee scheme
relationships).

All basic fee types that are not required for AGFS scheme 10 will not
include its associated role, meaning they won't be available for
selection in the graduated fees page.

Also:

- Refactor seed file for fee types, firstly to ensure it can be run
multiple times and only creates the seeded records if they don't exist
already. If they exist, update its attributes instead. Used a more OO
approach as well, as of now all the code was defined in a single file.