Fix caching issue with basic_fees association
The initialise_basic_fees is being called every single time an instance
of the claim is created in memory! (NOT very good! :( )

Due to the above, and the way the new basic fees were being added to the
claim, every single instance of the basic fee being added had no
awareness that another instance with the same fee type code had been
already added to the collection. Using the Rails CONVENTIONS (they exist
for a reason) `collection.build(attributes)` ensures that doesn't happen!