Add missing index to claims on uuid column
[This](https://github.com/ministryofjustice/Claim-for-Crown-Court-Defence/blob/fefa0f40b261c97b745044eb8d683685dd828231/db/migrate/20150804100037_add_uuids.rb#L5)
does NOT add ANY index to the table.

There's a misconception about the use of the Rails migration method
`add_column`: `index: true` is ignored

More info here:
https://makandracards.com/makandra/32353-psa-index-true-in-rails-migrations-does-not-work-as-you-d-expect

This adds the missing index as it's actually being used for quering
claims and without it becomes very unoptimized (having to do scan
through the whole table to find the required uuid).

Example of use here:

https://github.com/ministryofjustice/Claim-for-Crown-Court-Defence/blob/8b7412148c569587140e7d78903c53ced8e2e5d9/app/interfaces/api/v2/ccr_claim.rb#L8

NOTE: Creates the index concurrently not to impact the live service
given the claims table has quite a lot of records and non-concurrently
indexing would block the table until completed.