Fix caching issue
In Rails 5, if the record is not properly reloaded, it will use the
cached associations. In this specific case, one of the specs was
deleting the associated supplier number and then asserting that the
litigator claim was not valid, but the validation for the supplier
number in the claim was using a cached value for the associated supplier
numbers that the provider of the claim had.