Update the `roles` concern to prevent pluralising
The agfs_scheme_9 role was being pluralised to agfs_scheme_9s
internally.  As we are now going to expose that to vendors via
the API, it made sense to address it and ensure roles ending in
numerals are not pluralised.

Update existing references to agfs_scheme_9s and agfs_scheme_10s