Fix display of radio buttons for certification page
The certification page has two sets of radio buttons for the same
resource. Since Rails 5 `collection_radio_buttons` renders an extra
hidden field with an empty value for the given attribute
(more info here:
https://blog.bigbinary.com/2016/05/18/rails-5-add-a-hidden-field-on-collection-radio-buttons.html)

What that means is, even when we selected a value from one of the sets,
the other set was setting an empty value by default (which ended up
taking precedence).

This change removes the use of the hidden fields for this specific case.