Fix memory/db issue with basic fees on JSON import
With the existent code, the query to check if
the basic fee to be updated existed was performed directly to the
database.
Since upgraded to Rails 5, what I have asserted is the following
behaviour:

```ruby
[8] pry(API::Helpers::ApiHelper)> claim.basic_fees.pluck(:fee_type_id)
=> [1, 5, 6, 7, 8, 9, 11, 102]
[9] pry(API::Helpers::ApiHelper)> claim.fees.pluck(:fee_type_id)
   (0.6ms)  SELECT "fees"."fee_type_id" FROM "fees" LEFT OUTER JOIN
"fee_types" ON "fee_types"."id" = "fees"."fee_type_id" WHERE
"fees"."claim_id" = $1  [["claim_id", 69]]
=> [5, 6, 7, 8, 9, 11, 1]
```

The `basic_fees` method contains the new fee, whereas `fees` doesn't.

More:

```ruby
[37] pry(API::Helpers::ApiHelper)> claim.basic_fees.map(&:persisted?)
=> [true, true, true, true, true, true, true, false]
```

When inspecting if the basic fees where persisted, the one we were
interested in: 102, was not yet persisted in the database. What that
means, is that, with the previous implementation, the filter hit the
datbase directly where there was still no fee with fee_type = 102.

These changes perform that filter in memory and directly on the
basic_fees association to ensure we're getting the expected result.