DRY presentation of check your claim page
All the check your page claim was doing was a whole bunch of if/else
statements to display sections depending on what type of claim and what
kind of information it has.

These changes do 2 things:
- the sections that are required for a given type of claim are now
defined in its associated presenter and the view just iterates over
those sections.
- if a given section is required for a type of claim but its not to be
displayed given some data the claim has, that decision is taken on the
template for that section.