Add logic to claim "Check your claim" page to check invalid sections
The claim page will show 2 new components:
- A top level component that applies to the claim, which will be
displayed if the claim is not valid to be certified.
- A component per section, which will be displayed if that given section
requires some other sections to be filled before it can be completed.

Also:
- Refactored the logic to determine the validity of a claim/section.
Moved that logic to services given there was too much complexity to be
added directly to the claims model.

TODO: There's still quite a lot of DRY work to do, but these changes are
becoming to big to be able to address them:
- DRY summary pages. There's a lot of repetition in the summary page,
specially for the headers/status of the page