Prevent claim from being certified if they're not valid
Both new/create actions for the certifications controller were changed
to have a before hook to assert that the claim is valid and can be
certified.

If the claim is not valid, the user is redirected to the "Check Your
Claim" page and an error message is displayed.

Also:
- Removed duplicate check on new action for status of claim (was already
being done in a before hook)
- The new action used to validate the claim based on the current step
(at least there's step being step). With the changes introduced to the
validations to have them perform in isolation for each step, the
certifications would validate ONLY the default step (case details for
most of the claims). These changes guarantee that the validation is
performed on all the steps of the claim, which should be the only way to
proceed to the certification.