Assign totals for all claims before validation
The totals (fees, expenses, vat, etc) are now validated in the
appropriate steps where they're required.
Because in the first instance
the fees are not yet persisted during the validations, a 'total amount
claimed must be greater than 0' was being displayed. The reason for that
is the totals were only being updated after the records were saved.

These changes, ensure for non-API submissions (still need to check if
the API should follow the same behaviour or not) the totals are set
before validation so we can assert they're valid as expected.

Also:
- Refactored the calculations module to allow collections or single
records of fees to be provided for the totals calculation.