Basic fees are not displayed in CYC page if they're all blank
CYC - Check Your Claim page

The rationale behind this change has to do with the fact that these type
of fees are currently created blank before hand. To keep this behaviour
without impacting the rest of the application, the CYC page does not
display any basic fees if they're all blank.