Fixed fee page design implementation
#### What
Apply new design to fixed fees page

#### Ticket

[Pivotal](https://www.pivotaltracker.com/story/show/157116534)

#### Why
Simon asked nicely

#### How
CSS:
As we apply updated designs to the pages we will migrate the pages from V1 css files to V2. The migration is to simplify the CSS and pay off some tech dept. Some linting also apllied to the scss

note: during this migration their will be instances of duplication across both V1 and V2. This is as a result of shared views / form builders.

JS:
- minor config updates on Error handling module to work with updated semantics
- Refactor to remove nasty looking ternary statement

Ruby:
- `adp_text_field` updated semantics,  some styles and the specs
- `claims_helper` updated with css
- `base_fee` and `basic_fee` models updated
- Features updated as result of semantic changes

Views:
- implemented fixed fee design
- `dates_attended_fields_condensed` created to aid in migrating pages with shared templates
- using the `v2_stylesheets` value to control the loading of css and `basic` & `fixed` fee templates

GEMS:
- `gov_uk_date_fields` updated

Translations:
- updated to account for new structures in the views

--------